+++ view/client/src/main/java/net/eanfang/client/ui/activity/leave_post/bean/LeavePostDefaultRankingBean.java
@@ -14,7 +14,7 @@ public class LeavePostDefaultRankingBean {
     private String alertName;
     private String alertsCount;
     private int compareValue;
-    private int stationId;
+    private long stationId;
     private String date;
     /**
      * 0:年 1：季 2：月 3：周 4：天
